# docs/swagger.yaml
basePath: /v1
definitions:
  app_code.CodeType:
    enum:
    - success
    - user_not_found
    - password_error
    - service_error
    - auth_failed
    - wx_auth_failed
    - token_not_set
    - repeat
    - token_timeout
    - forbidden
    - user_exist
    - update_user_setting
    - get_user_setting
    - task_not_found
    - task_exist
    - bad_request
    - audit_param_invalid
    - not_found
    - task_run_not_found
    - create_task
    type: string
    x-enum-varnames:
    - Success
    - ErrorUserNotFound
    - ErrorUserPassword
    - ErrorServerError
    - ErrorAuthFailed
    - ErrorWxAuthFailed
    - ErrorTokenNotSet
    - ErrorRepeat
    - ErrorTokenTimeout
    - ErrorForbidden
    - ErrorUserExist
    - ErrorUpdateUserSetting
    - ErrorGetUserSetting
    - ErrorTaskNotFound
    - ErrorTaskExist
    - ErrorBadRequest
    - ErrorAuditParamInValid
    - ErrorNotFound
    - ErrorTaskRunNotFound
    - ErrorCreateTask
...
  entity.UserSetting:
    properties:
      address:
        type: string
      birthday:
        type: string
      education:
        type: string
      email:
        type: string
      emergency_contact:
        type: string
      emergency_phone:
        type: string
      emergency_relationship:
        type: string
      idcard:
        type: string
      insurance_end:
        type: string
      insurance_name:
        type: string
      insurance_photo:
        type: string
      insurance_start:
        type: string
      intro:
        type: string
      intro_user_id:
        type: string
      login_name:
        type: string
      married:
        type: string
      mingzu:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      sex:
        type: string
      user_id:
        type: string
      wechat_name:
        type: string
      work_category:
        type: string
    type: object
...
  http_util.Response:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/app_code.CodeType'
        example: success
      data: {}
      error:
        example: error message
        type: string
    type: object
...
  v1.doLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  v1.doLoginResponse:
    properties:
      role:
        type: string
      token:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
  v1.doMiniProgramLoginRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  v1.doMiniProgramLoginResponse:
    properties:
      phone:
        type: string
      token:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
  v1.doRegisterRequest:
    properties:
      password:
        type: string
      phone:
        type: string
      setting:
        $ref: '#/definitions/entity.UserSetting'
      username:
        type: string
    required:
    - password
    - username
    type: object
  v1.doRegisterResponse:
    properties:
      userId:
        type: integer
    type: object
host: task.mytool.zone
info:
  contact: {}
  description: Task Mini Program API
  title: Task Mini Program
  version: "1.0"
paths:
...
  /user/getSetting:
    get:
      consumes:
      - application/json
      description: 获取用户设置
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserSetting'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: 获取用户设置
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 登录参数
        in: body
        name: jsonBody
        required: true
        schema:
          $ref: '#/definitions/v1.doLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.doLoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: 登录
      tags:
      - 用户
  /user/miniProgramLogin:
    post:
      consumes:
      - application/json
      description: 小程序登录
      parameters:
      - description: 登录参数
        in: body
        name: jsonBody
        required: true
        schema:
          $ref: '#/definitions/v1.doMiniProgramLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.doMiniProgramLoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: 小程序登录
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 注册
      parameters:
      - description: 注册参数
        in: body
        name: jsonBody
        required: true
        schema:
          $ref: '#/definitions/v1.doRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.doRegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: 注册
      tags:
      - 用户
  /user/updateSetting:
    post:
      consumes:
      - application/json
      description: 更新用户设置
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户设置
        in: body
        name: jsonBody
        required: true
        schema:
          $ref: '#/definitions/entity.UserSetting'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: 更新用户设置
      tags:
      - 用户
swagger: "2.0"